fix: async errors when bucket needs to update the confirmation dialog.
Also fix regex for modes in parentheses
