+++ module/actor/actor.js
@@ -3063,12 +3063,10 @@ export class GurpsActor extends Actor {
         case 'm':
           refTags = taggedSettings.allAttackRolls.split(',').map(it => it.trim().toLowerCase())
           refTags = refTags.concat(taggedSettings.allMeleeRolls.split(',').map(it => it.trim().toLowerCase()))
-          isDamageRoll = true
           break
         case 'r':
           refTags = taggedSettings.allAttackRolls.split(',').map(it => it.trim().toLowerCase())
           refTags = refTags.concat(taggedSettings.allRangedRolls.split(',').map(it => it.trim().toLowerCase()))
-          isDamageRoll = true
           break
         case 'p':
           refTags = taggedSettings.allDefenseRolls.split(',').map(it => it.trim().toLowerCase())
@@ -3320,7 +3318,7 @@ export class GurpsActor extends Actor {
     switch (originType) {
       case 'attack':
         name = action.name.split('(')[0].trim()
-        mode = action.name.match(/\((.+?)\)/)?.[1]
+        mode = action.name.match(/\((.+)\)/)?.[1]
         const path = action.orig.toLowerCase().startsWith('m:') ? 'melee' : 'ranged'
         recurselist(this.system[path], (obj, _k, _d) => {
           if ((obj.originalName === name || obj.name === name) && (!mode || obj.mode === mode)) {
